% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{Building-class}
\alias{Building-class}
\title{Building Class}
\description{
An S4 class to represent building information, including various attributes
like street name, year of construction, energy usage, and more.
}
\section{Slots}{

\describe{
\item{\code{DEINR}}{numeric The building entrance number.}

\item{\code{STRNAME}}{character The street name.}

\item{\code{DPLZ4}}{numeric The postal code.}

\item{\code{EGID}}{numeric The unique building identifier.}

\item{\code{GKAT}}{numeric The building category.}

\item{\code{GKLAS}}{numeric The building class.}

\item{\code{GBAUJ}}{numeric The year of construction.}

\item{\code{GBAUP}}{numeric The period of construction.}

\item{\code{GABBJ}}{numeric The year of demolition.}

\item{\code{GAREA}}{numeric The surface area of the building.}

\item{\code{GASTW}}{numeric The number of floors.}

\item{\code{GEBF}}{numeric The energy relevant surface.}

\item{\code{GWAERZH1}}{character The heat generator for heating 1.}

\item{\code{GWAERZH2}}{character The heat generator for heating 2.}

\item{\code{GENH1}}{character The energy source for heating 1.}

\item{\code{GENH2}}{character The energy source for heating 2.}

\item{\code{GWAERSCEH1}}{character The information source for heating 1.}

\item{\code{GWAERSCEH2}}{character The information source for heating 2.}

\item{\code{GWAERDATH1}}{Date The revision date for heating 1.}

\item{\code{GWAERDATH2}}{Date The revision date for heating 2.}

\item{\code{GWAERZW1}}{character The heat generator for warm water 1.}

\item{\code{GWAERZW2}}{character The heat generator for warm water 2.}

\item{\code{GENW1}}{character The energy source for warm water 1.}

\item{\code{GENW2}}{character The energy source for warm water 2.}

\item{\code{GWAERSCEW1}}{character The information source for warm water 1.}

\item{\code{GWAERSCEW2}}{character The information source for warm water 2.}

\item{\code{GWAERDATW1}}{Date The revision date for warm water 1.}

\item{\code{GWAERDATW2}}{Date The revision date for warm water 2.}

\item{\code{utilisation_key}}{numeric The utilisation key according to SIA 380/1.}

\item{\code{climate_code}}{numeric The climate code of the closest climate station.}

\item{\code{energy_carrier}}{character The energy carrier according to SIA 380/1.}

\item{\code{walls_refurb_year}}{numeric The year of the last refurbishment of the walls.}

\item{\code{roof_refurb_year}}{numeric The year of the last refurbishment of the roof.}

\item{\code{windows_refurb_year}}{numeric The year of the last refurbishment of the windows.}

\item{\code{floor_refurb_year}}{numeric The year of the last refurbishment of the floor.}

\item{\code{heating_install_year}}{numeric The year of the installation of the heating system.}

\item{\code{heatEnergy}}{numeric Annual heat energy required per area, in MJ/m2 per year.}

\item{\code{emissionCoefficient}}{numeric Applied CO2 emission coefficient according to the given energy carrier, in kg/MJ.}

\item{\code{emissionsPerArea}}{numeric Annual CO2 emissions per area, in kg/m2 per year.}

\item{\code{emissionsTotal}}{numeric Total annual CO2 emissions, in kg per year.}

\item{\code{error_comments}}{character A string containing error messages.}
}}

