% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_regbl_db.R
\name{download_regbl_db}
\alias{download_regbl_db}
\title{Download and Prepare SQLite Database}
\usage{
download_regbl_db(db_filepath, overwrite = FALSE)
}
\arguments{
\item{db_filepath}{A character string specifying the file path where the SQLite database should be saved.}

\item{overwrite}{A logical value indicating whether to overwrite the existing database file if it already exists.}
}
\description{
This function downloads a ZIP file containing a SQLite database and related files from a specified URL.
It extracts the database file, moves it to a specified file path, and creates an index on the 'entrance' table.
}
\examples{
\dontrun{
# Example usage:
download_regbl_db("path/to/your/database.sqlite")
}
}
