% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egid_search.R
\name{egid_search}
\alias{egid_search}
\title{Retrieve EGID from Swiss GeoAdmin API Based on Address Information}
\usage{
egid_search(building)
}
\arguments{
\item{building}{A list containing optional address components: DPLZ4 (postal code), STRNAME (street name), and DEINR (house number).
These components are used to search the corresponding address.}
}
\value{
The input \code{building} list with an additional field \code{EGID}, which contains the unique identifier of the building.
}
\description{
This function queries the Swiss GeoAdmin API to find the EGID (unique building identifier)
based on provided address information. It constructs a search query using the building's
postal code (DPLZ4), street name (STRNAME), and house number (DEINR). The function retrieves
the most likely matching address and extracts the EGID from the result.
}
\examples{
building <- list(STRNAME = "Taubenstrasse 32 3011 Bern")
building <- egid_search(building)

}
