% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{euclidean_distance}
\alias{euclidean_distance}
\title{Euclidean Distance}
\usage{
euclidean_distance(x1, y1, x2, y2)
}
\arguments{
\item{x1}{The x-coordinate of the first point.}

\item{y1}{The y-coordinate of the first point.}

\item{x2}{The x-coordinate of the second point.}

\item{y2}{The y-coordinate of the second point.}
}
\value{
The Euclidean distance between the two points.
}
\description{
This function calculates the Euclidean distance between two points in a 2D plane.
}
