% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fill_building_data.R
\name{fill_building_data}
\alias{fill_building_data}
\title{Retrieve and Enhance Building Data with RegBl data and CO2 Emissions Calculations}
\usage{
fill_building_data(building)
}
\arguments{
\item{building}{A list or data frame containing at least minimal identifying information for a building,
which is then used to fetch and augment data from the RegBl database.}
}
\value{
An enhanced version of the input \code{building} data, including additional details and computed CO2 emissions.
}
\description{
This function enriches a given building's data record by fetching additional details from the RegBl API,
converting the data to the SIA format used by the co2calculatorPACTA2022 package, and computing the building's CO2 emissions. If the building
has been demolished, as indicated by a non-NA \code{GABBJ} value, the function halts and reports the demolition.
Otherwise, it proceeds to convert the format of the building data and calculate emissions using various
building attributes.
}
\examples{
building_data <- list(EGID = "123456")
enriched_building_data <- fill_building_data(building_data)
}
