% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_closest_station}
\alias{find_closest_station}
\title{Find the Closest Meteorological Station}
\usage{
find_closest_station(climate, building)
}
\arguments{
\item{climate}{A list of meteorological stations, where each station is an object
with properties \verb{$latitude}, \verb{$longitude}, and \verb{$code}. Latitude and longitude
should be numeric values representing the station's location.}

\item{building}{An object representing the building, with properties \verb{$DKODE}
for latitude and \verb{$DKODN} for longitude. Both should be numeric values.}
}
\value{
A character string representing the \verb{$code} of the closest meteorological
station to the given building.
}
\description{
This function calculates the Euclidean distance between a given building
and a list of meteorological stations, returning the code of the closest station.
}
