% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pkg_install}
\alias{pkg_install}
\title{pkg_install}
\usage{
pkg_install(packages, repos = "https://stat.ethz.ch/CRAN")
}
\arguments{
\item{packages}{Vector of packages to install.}

\item{repos}{Link to the CRAN repo to get the packages.}
}
\description{
Install one or more package from a given CRAN repository,
is they're not already installed.
}
\examples{
pkg_install(packages = c("knitr", "readxl"), repos = "https://stat.ethz.ch/CRAN")
}
