% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regbl_2_sia_converter.R
\name{regbl_2_sia_converter}
\alias{regbl_2_sia_converter}
\title{Convert Building Data from RegBl to SIA Format}
\usage{
regbl_2_sia_converter(building)
}
\arguments{
\item{building}{A named list or dataframe containing building data in RegBl format.}
}
\value{
The same building data structure but with additional values converted to the SIA format.
}
\description{
This function converts building data retrieved from the RegBl database into the format
required by the SIA (Swiss Society of Engineers and Architects) standards, specifically
for use with the co2calculatorPACTA2022 package
(see \url{https://maximecharriere.github.io/co2calculatorPACTA2022/}).
}
\details{
The conversion rules applied by this function include:
\itemize{
\item The energy relevant area is taken from \code{GEBF} if available, or calculated from \code{GAREA} times \code{GASTW}.
\item The construction year is taken from \code{GBAUJ} if available, or the earliest year given by \code{GBAUP}.
\item The closest climate station code is determined for the given building location.
\item The utilisation key is mapped from the RegBl building class (\code{GKLAS}) using a lookup table.
\item The energy carrier is mapped from the RegBl energy source type (\code{GENH1}) using a lookup table.
\item The number of floors (\code{floors}) and the year the heating system was installed
(\code{heating_install_year}) are directly assigned.
}
}
