% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_regbl.R
\name{request_regbl}
\alias{request_regbl}
\title{Request building data from the RegBl MADD API}
\usage{
request_regbl(building)
}
\arguments{
\item{building}{A list with the following elements:
\itemize{
\item DEINR: The building entrance number
\item STRNAME: The street name
\item DPLZ4: The postal code
}}
}
\value{
A list with the following elements:
\itemize{
\item EGID: The unique building identifier
\item GKAT: The building category
\item GBAUJ: The year of construction
\item GBAUP: The period of construction
\item GABBJ: The year of demolition
\item GAREA: The surface area of the building
\item GASTW: The number of floors
\item GEBF: The energy relevant surface
\item GWAERZH1-2: The heat generator for heating (1st and 2nd system)
\item GENH1-2: The energy source for heating
\item GWAERSCEH1-2: The information source for heating
\item GWAERDATH1-2: The revision date for heating
\item GWAERZW1-2: The heat generator for warm water
\item GENW1-2: The energy source for warm water
\item GWAERSCEW1-2: The information source for warm water
\item GWAERDATW1-2: The revision date for warm water
}
}
\description{
This function sends a request to the RegBl MADD API to get building data.
}
