% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_regbl_data.R
\name{request_regbl_data}
\alias{request_regbl_data}
\title{Request Building Information from Local SQLite Database}
\usage{
request_regbl_data(building, sqlite_conn = NULL)
}
\arguments{
\item{building}{A list or data frame representing the building, containing at least EGID, address components (DEINR, STRNAME, and DPLZ4), or parcel components (LPARZ, and DPLZ4).}

\item{sqlite_conn}{A SQLiteConnection object representing the database connection. This must be established prior to calling this function, for efficiency purpose.}
}
\value{
A modified version of the input building object, with additional information filled in from the SQLite database.
If no unique building is found, or if an error occurs during the query, the function stops with an error message.
}
\description{
This function retrieves detailed information about a building from a local SQLite database.
The request can be made using either the building's unique identifier (EGID), its address components
(DEINR, STRNAME, DPLZ4), or its parcel components (LPARZ, DPLZ4). The function queries the database,
processes the results, and fills in the building information.
}
\details{
The function constructs SQL queries to fetch data from the relevant tables in the SQLite database. After retrieving
the data, it processes the results to populate the building object with the appropriate types for each field, based
on metadata provided in \code{.constants$buildings_df_columns}.

The function assumes that \code{.constants$buildings_df_columns} is defined and contains metadata about the expected building data fields, including their types.

To download the SQLite database file, please visit https://www.housing-stat.ch/fr/madd/public.html, or run the \code{download_regbl_db()} function.
}
\examples{
\dontrun{
conn <- DBI::dbConnect(RSQLite::SQLite(), "path/to/database.sqlite")
building <- list(EGID = 123456)
updated_building <- request_regbl_sqlite(building, sqlite_conn = conn)
DBI::dbDisconnect(conn)
}

}
