% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{split_addresses}
\alias{split_addresses}
\title{Split Addresses into Street Name and House Number in a Data Frame}
\usage{
split_addresses(data, col_name = "Addresses")
}
\arguments{
\item{data}{A data frame containing a column with addresses.}

\item{col_name}{The name of the column containing the addresses.}
}
\value{
A data frame with the original addresses and two new columns: $street containing the street name, and $number containing the house number.
}
\description{
This function takes a data frame with a column containing addresses and splits each address into two components: the street name and the house number. The street is everything before the last sequence of digits which may include letters (e.g., "54A" or "10A").
}
