% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardize_buildings_df}
\alias{standardize_buildings_df}
\title{Standardize the Structure of a Building Data Frame}
\usage{
standardize_buildings_df(buildings_df, buildings_df_columns_names)
}
\arguments{
\item{buildings_df}{A data frame containing building information that needs to be standardized.}

\item{buildings_df_columns_names}{A character vector specifying the desired set of column names
that should be present in the buildings data frame.}
}
\value{
The input data frame augmented with any missing columns from the specified set,
with NA values in these new columns.
}
\description{
This function ensures that a given building data frame contains a specific set of columns.
It checks the data frame against a provided list of column names and adds any that are missing,
initializing them with NA values. This standardization is essential for consistent data handling
and analysis, particularly when integrating multiple building data sets with potentially varying structures.
}
\examples{
buildings_df <- data.frame(EGID = c(12345, 67890), STRNAME = c("Main St", "Second St"))
desired_columns <- c("EGID", "STRNAME", "DEINR", "DPLZ4")
standardized_df <- standardize_buildings_df(buildings_df, desired_columns)
# standardized_df will now include DEINR and DPLZ4 columns with NA values.
}
