% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_regbl_db.R
\name{update_regbl_db}
\alias{update_regbl_db}
\title{Download and Prepare SQLite Database}
\usage{
update_regbl_db(db_filepath)
}
\arguments{
\item{db_filepath}{A character string specifying the file path where the SQLite database should be saved.}
}
\description{
This function downloads a ZIP file containing a SQLite database and related files from a specified URL.
It extracts the database file, moves it to a specified file path, and creates an index on the 'entrance' table.
}
\examples{
\dontrun{
# Example usage:
update_regbl_db("path/to/your/database.sqlite")
}
}
