% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wgs84_to_lv95}
\alias{wgs84_to_lv95}
\title{Call Geodesy API to Convert WGS84 to LV95 Coordinates}
\usage{
wgs84_to_lv95(easting, northing, altitude)
}
\arguments{
\item{easting}{Ellipsoidal longitude in decimal degrees \link{°} on WGS84.}

\item{northing}{Ellipsoidal latitude in decimal degrees \link{°} on WGS84.}

\item{altitude}{Ellipsoidal height in meters \link{m} on WGS84. If not required, set to \code{NULL}.}
}
\value{
A list containing the easting, northing, and altitude in LV95 coordinate system.
If the altitude is specified, it returns the ellipsoidal height on Bessel.
}
\description{
This function sends a request to the Geodesy API to convert coordinates from WGS84 to LV95.
}
\references{
Geodesy API documentation: https://www.swisstopo.admin.ch/en/rest-api-geoservices-reframe-web
}
